TG-05: added corrections for status messages and postman collection endpoints

+++ src/controllers/projects.js
@@ -1,6 +1,14 @@
+import mongoose from 'mongoose';
 import Projects from '../models/Projects';
 
 const updateProject = async (req, res) => {
+  if (req.params.id && !mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(404).json({
+      message: `Project with id ${req.params.id} not found`,
+      data: undefined,
+      error: true,
+    });
+  }
   try {
     const { id } = req.params;
     const result = await Projects.findByIdAndUpdate(
@@ -8,32 +16,52 @@ const updateProject = async (req, res) => {
       { ...req.body },
       { new: true },
     );
-
-    return res.status(200).json({
-      message: 'The project has been updated',
-      data: result,
-      error: false,
+    if (result) {
+      return res.status(200).json({
+        message: `Project with id ${req.params.id} updated successfully`,
+        data: result,
+        error: false,
+      });
+    }
+    return res.status(404).json({
+      message: `Project with id ${req.params.id} not found`,
+      data: undefined,
+      error: true,
     });
   } catch (error) {
-    return res.json({
-      message: `There has been an error: ${error}`,
+    return res.status(500).json({
+      message: `Server error: ${error}`,
       error: true,
     });
   }
 };
 
 const deleteProject = async (req, res) => {
+  if (req.params.id && !mongoose.Types.ObjectId.isValid(req.params.id)) {
+    return res.status(404).json({
+      message: `Project with id ${req.params.id} not found`,
+      data: undefined,
+      error: true,
+    });
+  }
   try {
     const { id } = req.params;
     const result = await Projects.findByIdAndDelete(id);
-    return res.status(200).json({
-      message: 'The project has been deleted',
-      data: result,
-      error: false,
+    if (result) {
+      return res.status(200).json({
+        message: 'Project deleted successfully',
+        data: result,
+        error: false,
+      });
+    }
+    return res.status(404).json({
+      message: `Project with id ${req.params.id} not found`,
+      data: undefined,
+      error: true,
     });
   } catch (error) {
-    return res.json({
-      message: `There has been an error: ${error}`,
+    return res.status(500).json({
+      message: `Server error: ${error}`,
       error: true,
     });
   }
